chatbot_backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]>=0.22,<1.0
motor==3.7.1
numpy==2.2.6
openai==1.82.0
pydantic==2.11.5
pymongo==4.13.0
python-dotenv==1.1.0
requests==2.32.4
scikit_learn==1.7.0
xgboost==3.0.2
pandas==2.3.0
python-multipart==0.0.20